requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
gunicorn==21.2.0
idna==3.4
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
whitenoise==6.5.0
